requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.4.0
bidict==0.23.1
blinker==1.9.0
cachelib==0.13.0
certifi==2024.7.4
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.2
Flask==2.3.2
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-Session==0.8.0
Flask-SocketIO==5.3.4
Flask-SQLAlchemy==3.0.3
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_metadata==8.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
Mako==1.3.8
MarkupSafe==2.1.5
msgspec==0.19.0
openai==1.40.3
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
python-engineio==4.11.1
python-socketio==5.12.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.16
tqdm==4.66.5
typing_extensions==4.12.2
Werkzeug==2.3.3
wsproto==1.2.0
zipp==3.19.2
gunicorn==23.0.0

